requirements.txt
flask
python-dotenv
langchain
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
huggingface_hub
transformers
torch
pypdf
langchain_openai
langchain_core
requests
flask_cors
flask_sqlalchemy
flask_migrate
flask_wtf
flask_bootstrap
langchain_groq
langchain_huggingface
langchain_text_splitters
langchain_document_loaders
langchain_vectorstores
langchain_chains
